requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
langchain>=0.1.0
pypdf>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.28.0
langchain_community